+++ src/main/java/com/medallia/data/Segment.java
@@ -1,9 +1,9 @@
 package com.medallia.data;
 
 public class Segment {
-	public final long[][] rawData;
+	public final int[][] rawData;
 
 	public Segment(int cols) {
-		rawData = new long[cols][];
+		rawData = new int[cols][];
 	}
 }
